+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Using string insted of using an object when computing the hash from the theme
 
 ## [0.6.6] - 2024-10-15
 - Exposed gradient background image as a css variable